ref: move SerialPort code into its own file
